# Merge storages into archetypes for faster access