Brings 5-10% speedup of archetype access. ~Not sure if this really justifies the added code.~

New and old version alternating for better comparability over fluctuating machine load.

```
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
BenchmarkArchetypeAccess1_1000-8         1618312               717.6 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccessOld1_1000-8      1581909               753.6 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccess2_1000-8         1693066               709.0 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccessOld2_1000-8      1615272               843.7 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccess3_1000-8         1668674               706.7 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccessOld3_1000-8      1615597               749.5 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccess4_1000-8         1685044               713.2 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccessOld4_1000-8      1624300               738.7 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccess_1000-8          1682244               720.4 ns/op             0 B/op          0 allocs/op
BenchmarkArchetypeAccessOld_1000-8       1613794               743.7 ns/op             0 B/op          0 allocs/op
```